:recycle: Refactor configuration handling by introducing config.interface and load-env-variables for improved structure and clarity
