+++ functions/src/app/config/config.interface.ts
@@ -1,0 +1,4 @@
+import { EnvVariables } from './load-env-variables'
+import { Secrets } from './secrets.provider'
+
+export type Configuration = EnvVariables & Secrets
