+++ functions/src/app/html-to-pdf/pdf-generators/pdf-generator.adapter.ts
@@ -1,7 +1,7 @@
 import { Buffer } from 'node:buffer'
 import { ConfigService } from '@nestjs/config'
 import { PDFOptions } from 'puppeteer-core'
-import { Configuration } from '../../config/configuration'
+import { Configuration } from '../../config/config.interface'
 
 export type PdfGenerationOptions = Pick<PDFOptions, 'format' | 'landscape' | 'displayHeaderFooter'>
 export abstract class PdfGenerator {
