+++ functions/src/app/secret-manager/secret-manager.service.ts
@@ -2,7 +2,7 @@ import { Buffer } from 'node:buffer'
 import { SecretManagerServiceClient } from '@google-cloud/secret-manager'
 import { Injectable, Logger } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
-import { Configuration } from '../config/configuration'
+import { Configuration } from '../config/config.interface'
 
 @Injectable()
 export class SecretManagerService {
